Cache `docker` image
This seems to be the last big thing for caching. We seem to be
rebuilding this image each time in CI. That's not the worst thing
(because `docker` caching is actually more complex than most people
would admit) but it does mean that we're spending a lot of time building
effectively the same thing.

We opt to make a cache of the image tar in a local directory, and use
that in CI. It's possible that we're going to regret this decision later
when we have troubles invalidating the cache. But we'll roll with this
for a bit, and see how it goes.

For posterity, the first commit in this change clocked in at 2m 17s; the
second commit with the cache hit clocked in at 1m 56s. A savings of 15%!

Branch: joneshf/cache-docker-image
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/86