Add a helper that returns a provider block string
Most of the time, we don't actually care about the connection
credentials for a provider, we just want one that works. Instead of
putting the onus on the caller to figure this out each time, we give
back a provider block string. This helps with the Resource acceptance
test helper case.

We also make a helper that returns a `lucirpc.Client` along with a
provider block string. This helps with the Data Source acceptance test
case.

Branch: joneshf/add-a-helper-that-returns-a-provider-block-string
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/90