Re-order `system.system` resource acceptance test
Since the `system.system` resource seems to always need to exist in
order for an OpenWrt server to run, and because there can be only one
`system.system`, we shouldn't try to emulate the use case of creating
one (as nobody would create a `system.system`).  Instead, we assume one
is already there. We start by importing it, and going from there.