Provide a helper for Terraform test steps
We're doing roughly the same thing everywhere. We also have to make sure
we get the provider setup right, as otherwise it'll make the test fail.
In order to make things a bit easier, we provide a helper that bundles
up a bunch of that logic.

Branch: joneshf/provide-a-helper-for-terraform-test-steps
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/94