Track the `nixos-22.11` branch in the flake
To try and keep things up to date easier, we're going to switch to using
a reference rather than a revision. It *seems* like a reference can be
updated more easily, but it's not actually clear how this will work out
in practice. And the `nix` docs don't really explain it too well:
https://nixos.org/manual/nix/stable/command-ref/new-cli/nix3-flake#flake-references.
The explanation on Zero to Nix seems to imply that it works the way
we're thinking, though:
https://zero-to-nix.com/concepts/flakes#lockfile.

As usually, we should be able to back out of this change if it proves to
be a bad decision.