Undo changes to server mocks
We got a little carried away, and changed the types of the server mocks
in `lucirpc/client_test.go`. These `body`s are supposed to be plain
`map[string]json.RawMessage`s because they represent the underlying data
being sent to LuCI's JSON-RPC API. They don't represent the (slightly)
higher level interface of the options for a section.

We swap these back so we don't end up with hard to diagnose errors once
we add more behavior for section options.