Type the LuCI JSON-RPC options
We want a little more help doing the right thing here. We make a case
for each of the different types of options, and use that as the
interface for LuCI's JSON-RPC options.

This change makes it so the logic of how to deal with UCI's
representation of data types stays closer to the LuCI JSON-RPC API. We
probably should have done this from the beginning. Better late than
never, though.

Branch: joneshf/type-the-luci-json-rpc-options
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/106