Add support for `network.interface`
Much like the `network.device`, we add support for both the Data Source
and Resource. Since `network.device` is so complex, we're only
implementing a subset of it.

Also like the `network.device`, the documentation is less than ideal.
It's separated into multiple pages. And each page is slightly different.
Maybe there's just a different way to find all of this information?

Branch: joneshf/add-support-for-network-interface
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/109