Add a `make` target to release a new patch version
We want to make things a little easier to deal with. So we add a `make`
target that lets us release a new patch version easily.

Since patch versions are the only bumping we've been doing, we only add
support for bumping patch versions. We can add support for other
versions when we get to that point.

Branch: joneshf/add-a-make-target-to-release-a-new-patch-version
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/111