Split each "resource" to its own file
We're using the term "resource" to mean both Data Source and Resource.
As it seems like the only real difference between a Data Source and
Resource is that a Resource has CRUD, while a Data Source only has R.

Anyway, we're splitting these because it's actually becoming a bit much
to have all of the resources together when running the tests. A simple
change to one of the resoruces in the `network` package means that all
of the acceptance tests have to run again. We'd like it to be less
painful. With the packages split, a change to one resource shouldn't
affect the others.

We also move the network stuff from being in an "internal" package to
being a public package. We don't really care if things break
right now. We're still `0.0.x`, so we shouldn't care so much about
structure at the moment.

The diff looks worse than it actually is. It's largely just moving files
and removing stutter.

Branch: joneshf/split-each-resource-to-its-own-file
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/113