Add support for `network.switch_vlan`
Continuing with the legacy support for `swconfig`, we want to be able to
support VLANs. These use the `network.switch_vlan` config. So, we add
support for this.

Again, this is pretty straight-forward config. So it's a relatively low
lift.

Branch: joneshf/add-support-for-network-switch-vlan
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/115