Coerce booleans and integers to other types
Since UCI doesn't really have types anywhere, we have to be very careful
about how we parse things. If an option has a value of `"1"`, this could
be a boolean, integer, or string; depending on the config it's a part
of. Frankly, this is an incredibly bad representation. And it feels like
the sort of thing where a language is not designed, but extended over
years without taking a step back and removing these awkward cases.

We add support for coercing the different types that we can at least
identify as possibly being something else. We almost surely are missing
some cases here. But, we'll deal with that when they arise. For now, we
at least can coerce some booleans and integers to other types.
Hopefully, this catches most of the cases.

Branch: joneshf/coerce-booleans-and-integers-to-other-types
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/116