Move the other attributes into OpenWrt struct
We had a weird split between stuff that was global to the
`acceptancetest` package, and local to the OpenWrt server. Instead of
keeping the two separate, we move all of those other attributes into the
OpenWrt server struct. It's irrelevant that they're always the same
value. The important part is that it's consistent.