Add a `OpenWrtServer.ProviderBlock` method
This doesn't really change much for the current usages, but trying to
make sure we use the correct provider setup is a little hard when it's
locked in the helpers file. We add a method on the `OpenWrtServer` that
creates a provider block.

This should make it easier to use the server for other uses.