Add support for `wireless.wifi-device`
We move on to the wireless stuff. First up is `wireless.wifi-device`. We
only handle a small subset of the options, since it's pretty extensive
what all this config can do. We'll have to expand this later. But, it
should be enough for simpler purposes.

Branch: joneshf/add-support-for-wireless-wifi-device
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/121