Fix UCI type for `wireless.wifi-device`
This typo is everything that's wrong with UCI. There's no consistency in
the naming scheme. And, there's nothing stopping you from using the
incorrect UCI type.

If there's a way to accurately test that we're using the correct thing,
we should switch to it. But, it's not clear how we'd actually catch a
mistake like this.

In any case, we fix this typo, so a `wireless.wifi-device` can actually
be created now.