Fix UCI type for `wireless.wifi-device`
This typo is everything that's wrong with UCI. There's no consistency in
the naming scheme. And, there's nothing stopping you from using the
incorrect UCI type. Nothing failed for any of the tests here because at
its core, UCI allows basically any config to have any name with any
section with any options. It's way too lenient, and it means that you
have to be extra careful whenever you're using it directly (like we're
doing here).

If there's a way to accurately test that we're using the correct thing,
we should switch to it. But, it's not clear how we'd actually catch a
mistake like this.

In any case, we fix this typo, so a `wireless.wifi-device` can actually
be created now.

Branch: joneshf/fix-uci-type-for-wireless-wifi-device
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/124