Add support for `wireless.wifi-iface`
We make it possible to define wireless networks. This is maybe the last
big piece to having a usable OpenWrt router in the 21st century. Most
everyone is going to have a need for at least one wireless network these
days (cell phones and laptops most notably).

This is much like the rest of the resources. Because wireless networks
have a lot of features, we only support a subset to start. We'll add
more as we go along. But, this should be enough to get the minimum
going.