Remove a lot of parallelism in the tests
The tests have gotten incredibly flaky over time. It's to the point that
it took eight attempts of the test job on the `main` branch to pass:
https://github.com/joneshf/terraform-provider-openwrt/actions/runs/4592301778.
That's too unreliable to be useful.

It's been pretty hard to diagnose what the issue is, but the best lead
we've got is that we're trying to run too many servers all at once, and
it's causing them to behave poorly. Since we're taking an educated guess
at the problem, the solution is also an educated guess: limit the
parallelism in the tests. This might not actually solve the issue for
us, but limited testing is showing that the tests succeed more often
than before.

If this doesn't work, we can always try something else. But we need to
start somewhere, and this seems like the most promising fix to make.