Add support for `dhcp.dnsmasq`
Moving on to DHCP stuff, we want to be able to configure the Dnsmasq
stuff. As with all of the other resources we add, this is only a subset
of what's available. We'll flesh out more as time goes on.

Branch: joneshf/add-support-for-dhcp-dnsmasq
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/127