Add support for `dhcp.odhcpd`
This seems to be one of the few resources which is not configurable
through LuCI. We still do want to support it, though. So we add support
for it.

It seems like we've added all the options. But if there are any that are
missing, we can add them later.