Fix `dhcp.dhcp` required fields
As it turns out, there's actually a case where these "required" fields
don't exist from the get-go. For WAN interfaces, they already have a
DHCP pool setup that is ignored. There's no `leasetime`, `limit`, or
`start` option associated with them. So you end up in this weird state
of questioning whether you should add these three fields to satiate this
Terraform provider, not import the WAN DHCP pool, delete the DHCP pool
from UCI all up, or something equally obtuse.

None of these things should be a thing someone has to think about. So we
change the attributes to not be required all the time, but only when the
`ignore` attribute is not `true`. We do this by adding a new validator
that checks the other attribute for its value and acts accordingly. We
update the tests to validate this behavior.

This is yet another example of how not having sum types causes so much
extra complexity. Ignoring the fact that OpenWrt shouldn't ship with an
ignored DHCP pool on the WAN interface, this situation is trivial (in
implementation) to represent with a sum type. But trying to represent it
with nothing by product and exponent types is incredibly complex. Ah
well, maybe one day things will get better…