Add support for `dhcp.host`
We want to be able to statically assign an IP address to a known MAC
address. This `dhcp.host` resource gives us exactly that. It's a fairly
easy going resource, so not much here.

Like the other resources we've added, we only add the minimum to make it
useful. We'll come back and add more in the future.