Add support for `dhcp.domain`
Last resource for the `dhcp` config. This is a bit redundant with the
`dhcp.host` stuff. But, it has a purpose. There are times when you need
to rebind a host name to an IP address (like when you want to provide a
subdomain). And this is where the `dhcp.domain` is useful.

It's unclear if there is more to this resource. There doesn't seem to be
any documentation on the OpenWrt wiki. There's only two fields in the
LuCI UI, so it seems like the whole thing has been implemented. If not,
we can always add it later.

Branch: joneshf/add-support-for-dhcp-domain
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/132