Include go files in go test cache key
We want the caches to update whenever there are changes to the files.
This way, when we alter any of the tests, they should update the cache
with their latest result. We do this by appending a hash of the go
files to the cache key.

We still want to fallback to the nearest cache that works though, so we
can bring it any test results that might be successfully cached, and
only worry about the new stuff. We do this by making sure one of the
restore keys drops the appended hash.